Add files via upload
This code contains a working version of the python game using the turtle library. I did not write this code. This code contains the basis for how my code will work, but there are other things I will need to implement in later iterations of the project that will make my code significantly different from this code. The next step for me in this project is to implement a way for the python game to play over a .jpg file. This way, when the other programs are done, they will work together without errors.