Add files via upload
Heyo, this is the snake game that I was not able to test because the window would not pop up on my computer. right now, this code should just open a window. If it works for you when you integrate it, i will keep writing it, but if not, i will try to make the turtle version of it work better.